llm_helper_utils/vllm_requirements.txt
accelerate~=0.20.3
aiohttp-cors~=0.7.0
amrlib~=0.7.1
apache-beam>=2.70.0
black~=22.3.0
cmake>=3.26.4,<4.0
codecarbon~=1.2.0
colorful~=0.5.6
commonmark~=0.9.1
cupy-cuda12x==13.4.1
datasets>=4.5.0,<5.0
decord==0.6.0
email_validator>=2.2.0,<3.0
evaluate~=0.4.0
faiss-cpu==1.7.4
fastapi-cli>=0.0.7,<0.1
flax==0.5.3
fugashi~=1.2.1
gcsfs>=2023.10.0,<2024.0
hf-doc-builder>=0.4.0,<0.5
hf_transfer>=0.1.9,<0.2
httptools>=0.6.4,<1.0
ipadic~=1.0.0
jupyterthemes~=0.20.0
kenlm==0.1
keras-nlp==0.5.2
librosa>=0.10.0.post2,<0.11
lit==16.0.6
nvidia-cufft-cu11==10.9.0.58
nvidia-ml-py>=12.570.86,<13.0
onnxconverter-common>=1.13.0,<2.0
opencensus>=0.11.4,<0.12
opentelemetry-exporter-otlp>=1.34.1,<2.0
opentelemetry-semantic-conventions-ai>=0.4.9,<0.5
optuna>=3.2.0,<4.0
outlines~=0.1.11
parameterized~=0.9.0
phonemizer>=3.2.1,<4.0
pinecone-client~=2.2.1
pip>=25.3
polars==0.16.18
pomegranate==0.15.0
py-spy>=0.4.0,<0.5
pyarrow-hotfix~=0.6
pybind11>=2.13.6,<3.0
pyctcdecode>=0.5.0,<0.6
pytest-timeout~=2.1.0
pytest-xdist~=3.3.1
python-dotenv>=1.0.1,<2.0
python-multipart>=0.0.20,<0.1
requests-toolbelt~=0.10.1
rhoknp>=1.3.2,<2.0
rjieba~=0.1.11
rouge-score~=0.1.2
s3fs>=2023.10.0,<2024.0
sacrebleu~=1.5.1
sacremoses~=0.0.53
sentence-transformers==2.2.2
sigopt~=8.8.0
SudachiDict-core==20230110
tensorboard-plugin-wit~=1.8.1
tensorboardX>=2.6,<3.0
tensorflow-estimator==2.11.0
tf2onnx>=1.14.0,<2.0
timeout-decorator~=0.5.0
timm~=0.9.2
unidic~=1.1.0
unidic-lite~=1.0.8
uvloop>=0.21.0,<0.22
virtualenv>=20.29.1,<21.0
vllm==0.11.0
websockets>=14.2,<15.0
wheel>=0.45.1,<1.0
word2number~=1.1